Update app.py flask dynamo - socket
[WARNING]	2024-01-21T13:31:43.679Z	2d10edde-4172-444e-ab9b-711a707d07fd	Exception raised while handling cache file '/var/task/flask_session/1818af657c5eabe02273a2d491f036d1'
Traceback (most recent call last):
  File "/var/task/cachelib/file.py", line 248, in set
    fd, tmp = tempfile.mkstemp(
  File "/var/lang/lib/python3.9/tempfile.py", line 336, in mkstemp
    return _mkstemp_inner(dir, prefix, suffix, flags, output_type)
  File "/var/lang/lib/python3.9/tempfile.py", line 255, in _mkstemp_inner
    fd = _os.open(file, flags, 0o600)
OSError: [Errno 30] Read-only file system: '/var/task/flask_session/tmp47wx01x3.__wz_cache'



The error indicates that Flask is encountering an issue when trying to write to a temporary file in the /var/task/flask_session/ directory. The underlying cause appears to be that the file system is read-only in the AWS Lambda environment.

AWS Lambda functions run in an environment where the file system is generally read-only. Lambda functions are designed to be stateless, and any temporary storage or caching should be done in other ways. Writing to the local file system, especially in read-only areas, is not recommended in a serverless environment like Lambda.

For Flask applications running on AWS Lambda, it's common to use other mechanisms for session storage. One approach is to use an external storage service, such as DynamoDB, to store session data. Another option is to use a serverless session manager that stores session data outside the Lambda function, allowing for scalability and avoiding file system issues.